+++ app/src/test/java/com/novacut/editor/DependencyFreshnessTest.kt
@@ -33,6 +33,7 @@ class DependencyFreshnessTest {
         "navigation",
         "coil",
         "okhttp",
+        "sqlite",
         "lottieCompose",
         "onnxruntime",
         "mediapipe",